feat: only write manifest if it changes

+++ src/cargo/scripts/manifest.ts
@@ -5,8 +5,9 @@ import { EOL } from "https://deno.land/std@0.173.0/fs/mod.ts";
 type Script = [string, string];
 
 type ScriptManifestConfig = {
-  path: string;
   manifestPath: string;
+  manifestFileName: string;
+  manifestAssetspath: string;
   entryPoints: EntryPoints;
   isProd?: boolean;
 };
@@ -15,32 +16,58 @@ export async function scriptsManifest(
   config: ScriptManifestConfig,
 ) {
   const _scripts: Script[] = [];
-  const scriptAssetPath = join(config.manifestPath, config.path);
+  const scriptManifestAssetPath = join(
+    config.manifestPath,
+    config.manifestAssetspath,
+  );
+  const scriptManifestPath = join(config.manifestPath, config.manifestFileName);
 
-  await Deno.mkdir(scriptAssetPath, { recursive: true });
+  await Deno.mkdir(scriptManifestAssetPath, { recursive: true });
 
   const _files = await bundle(
     config.entryPoints,
   );
 
   for (const [name, content] of _files) {
-    _scripts.push([name, join(scriptAssetPath, name)]);
+    _scripts.push([name, join(scriptManifestAssetPath, name)]);
     await Deno.writeFile(
-      join(scriptAssetPath, name),
+      join(scriptManifestAssetPath, name),
       content,
     );
   }
 
-  await Deno.writeTextFile(
-    join(config.manifestPath, ".scripts.ts"),
-    scripts(_scripts),
-  );
+  await writeManifest(_scripts, scriptManifestPath);
 }
 
 function scripts(scripts: Script[]) {
   return `export default [${
     scripts.map(([name]) => `"${name}"`).join(", ")
   }];${EOL.LF}`;
+}
+
+async function writeManifest(manifest: Script[], path: string) {
+  const _content = scripts(manifest);
+
+  try {
+    const existingManifest = await Deno.readTextFile(
+      path,
+    );
+
+    if (existingManifest !== _content) {
+      return await Deno.writeTextFile(
+        path,
+        _content,
+      );
+    }
+  } catch (e) {
+    if (e instanceof Deno.errors.NotFound) {
+      return await Deno.writeTextFile(
+        path,
+        _content,
+      );
+    }
+    throw e;
+  }
 }
 
 async function bundle(entryPoints: EntryPoints) {
